ENH: Overhaul
* Add TLS options to dicom send
* Bring up to poetry qa-ci spec
* reorganize code
* Use fw-file when we can
* Use fw-core-client when we can

+++ fw_gear_dicom_send/dicom_send.py
@@ -3,16 +3,18 @@
 import logging
 import os
 import shutil
+import sys
 import tarfile
 import zipfile
 from pathlib import Path
 
-import pydicom
-import flywheel
+import backoff
+from fw_file.dicom import DICOMCollection
+from fw_core_client import CoreClient, ClientError, ServerError
 
-from utils import tag_and_transmit
-from utils import report_generator
-
+from . import tag_and_transmit
+from . import report_generator
+from .parser import get_client, TLSOpt
 
 log = logging.getLogger(__name__)
 
@@ -27,79 +29,64 @@ def prepare_work_dir_contents(infile, work_dir):
         infile (pathlib.PosixPath): The absolute path to the input file.
         work_dir (pathlib.PosixPath): The absolute path to the working directory where
             the DICOM files are placed.
-
-    Returns:
-        None.
-
     """
     log.info("Arrange dicom-send input.")
 
     if zipfile.is_zipfile(infile):
-
+        log.info("Found input zipfile {infile}, unzipping")
         try:
-
             with zipfile.ZipFile(infile, "r") as zip_obj:
-                log.info(f"Establishing input as zip file: {infile}")
                 exit_if_archive_empty(zip_obj)
                 zip_obj.extractall(work_dir)
-
         except zipfile.BadZipFile:
-            log.exception(
-                (
-                    "Incorrect gear input. "
-                    "File is not a zip archive file (.zip). Exiting."
-                )
-            )
-            os.sys.exit(1)
+            log.exception("Input looks like a zip but is not valid")
+            sys.exit(1)
 
     elif tarfile.is_tarfile(infile):
-
+        log.info("Found input tarfile {infile}, untarring")
         try:
             with tarfile.open(infile, "r") as tar_obj:
-                log.info(f"Establishing input as tar file: {infile}")
                 exit_if_archive_empty(tar_obj)
                 tar_obj.extractall(work_dir)
-
         except tarfile.ReadError:
-            log.exception(
-                (
-                    "Incorrect gear input. "
-                    "File is not a compressed tar archive file (.tgz). Exiting."
-                )
-            )
-            os.sys.exit(1)
+            log.exception("Input looks like a tar but is not valid")
+            sys.exit(1)
 
     else:
         log.info(f"Establishing input as single DICOM file: {infile}")
-        try:
-            # If valid DICOM file, move to working directory; otherwise, exit
-            pydicom.filereader.dcmread(infile, force=True)
-            shutil.copy2(infile, work_dir)
-        except pydicom.errors.InvalidDicomError:
-            log.info("Input file is not a valid DICOM file. Exiting.")
-            os.sys.exit(1)
 
     log.info("Input for dicom-send prepared successfully.")
 
 
 def exit_if_archive_empty(archive_obj):
     """If the archive contents are empty, log an error and exit."""
-    if type(archive_obj) == zipfile.ZipFile:
-        size_contents = sum([zipinfo.file_size for zipinfo in archive_obj.filelist])
-
-    elif type(archive_obj) == tarfile.TarFile:
-        size_contents = sum([tarinfo.size for tarinfo in archive_obj.getmembers()])
-
+    size = 0
+    if isinstance(archive_obj, zipfile.ZipFile):
+        size = sum([zipinfo.file_size for zipinfo in archive_obj.filelist])
+    elif isinstance(archive_obj, tarfile.TarFile):
+        size = sum([tarinfo.size for tarinfo in archive_obj.getmembers()])
     else:
         log.info(
             "Unsupported archive format. Unable to establish size of input archive. "
             "Exiting."
         )
-        os.sys.exit(1)
+        sys.exit(1)
 
-    if size_contents == 0:
+    if size == 0:
         log.error("Incorrect gear input. Input archive is empty. Exiting.")
-        os.sys.exit(1)
+        sys.exit(1)
+
+def get_retry_time() -> int:
+    """Helper function to return retry time from env."""
+    return int(os.getenv("FLYWHEEL_DOWNLOAD_RETRY_TIME", "10"))
+
+
+@backoff.on_exception(backoff.expo, ServerError, max_time=get_retry_time)
+def download_file(fw: CoreClient, acq_id: str, file_name: str, dest: Path):
+    """Download file from acquisition with retry."""
+    resp = fw.get(f"/acquisitions/{acq_id}/files/{file_name}", stream=True)
+    with open(dest, 'wb') as fp:
+        fp.write(resp.content)
 
 
 def download_and_send(
@@ -109,6 +96,7 @@ def download_and_send(
     work_dir,
     destination,
     called_ae,
+    tls: TLSOpt,
     port=104,
     calling_ae="flywheel",
     group="0x0021",
@@ -128,6 +116,7 @@ def download_and_send(
             the DICOM files are placed.
         destination (str):The IP address or hostname of the destination DICOM server.
         called_ae (str):The Called AE title of the receiving DICOM server.
+        tls (TLSOpt): TLS options
         port (int) = Port number of the listening DICOM service.
         calling_ae (str): The Calling AE title.
         group (str): The DICOM tag group to use when applying tag to DICOM file.
@@ -136,68 +125,66 @@ def download_and_send(
 
     Returns:
         tuple:
-            DICOMS_PRESENT (int): The number of DICOM files for which transmission was attempted.
-            DICOMS_SENT (int): The number of DICOM files transmitted.
+            dcms_present (int): The number of DICOM files for which transmission was attempted.
+            dcms_sent (int): The number of DICOM files transmitted.
 
     """
     log.info("Downloading DICOM files.")
-    DATA_FLAG = False
-    DICOMS_SENT = 0
-    DICOMS_PRESENT = 0
+    dcms_sent = 0
+    dcms_present = 0
 
     # Create input directory if it doesn't exist
     if not Path(input_dir).is_dir():
         os.mkdir(input_dir)
 
     # Instantiate instance connection and load acquisitions in session
-    fw = flywheel.Client(api_key)
-    acquisitions = fw.get_session_acquisitions(session_id)
+    fw = get_client(api_key)
+    acquisitions = fw.get(f"/api/sessions/{session_id}/acquisitions")
 
     # In a session, the possible downloads include any combination of:
     # zip archive, tgz archive, a single DICOM file, multiple DICOM files, or empty.
     # All cases are handled by parsing the downloaded files in the input directory
     # and calling downstream functions accordingly.
     for acq in acquisitions:
+        log.debug(f"Processing acquisition {acq.label}")
         for file in acq.get("files"):
             if file.type == "dicom":
 
                 file_path = Path(f"{input_dir}/{file.name}")
-                fw.download_file_from_acquisition(acq.id, file.name, file_path)
-    
-                if file_path.is_file():
-                    # A file with file.type = dicom has been downloaded
-                    DATA_FLAG = True
+                download_file(fw, acq.id, file.name, file_path)
 
-                    dicoms_present, dicoms_sent = run(
+                if file_path.exists() and file_path.is_file():
+                    present, sent = run(
                         api_key,
                         acq.id,
                         file_path,
                         work_dir,
                         destination,
                         called_ae,
+                        tls,
                         port,
                         calling_ae,
                         group,
                         identifier,
                         tag_value,
                     )
-                    DICOMS_SENT += dicoms_sent
-                    DICOMS_PRESENT += dicoms_present
+                    dcms_sent += sent
+                    dcms_present += present
 
                     # Remove contents of working directory because we assume multiple
                     # downloaded files and need a clean workspace for each run.
                     shutil.rmtree(work_dir)
                     os.mkdir(work_dir)
 
-    if DATA_FLAG is False:
+    if not dcms_present:
         log.error(
             "No DICOM files were available for download for session with ID: "
             f"{session_id}. The file.type must be set to dicom for download "
             "to occur. Exiting."
         )
-        os.sys.exit(1)
+        sys.exit(1)
 
-    return DICOMS_PRESENT, DICOMS_SENT
+    return dcms_present, dcms_sent
 
 
 def run(
@@ -207,6 +194,7 @@ def run(
     work_dir,
     destination,
     called_ae,
+    tls: TLSOpt,
     port=104,
     calling_ae="flywheel",
     group="0x0021",
@@ -224,6 +212,7 @@ def run(
             the DICOM files are placed.
         destination (str):The IP address or hostname of the destination DICOM server.
         called_ae (str):The Called AE title of the receiving DICOM server.
+        tls (TLSOpt): TLS options
         port (int) = Port number of the listening DICOM service.
         calling_ae (str): The Calling AE title.
         group (str): The DICOM tag group to use when applying tag to DICOM file.
@@ -232,18 +221,14 @@ def run(
 
     Returns:
         tuple:
-            DICOMS_PRESENT (int): The number of DICOM files for which transmission was attempted.
-            DICOMS_SENT (int): The number of DICOM files transmitted.
+            dcms_present (int): The number of DICOM files for which transmission was attempted.
+            dcms_sent (int): The number of DICOM files transmitted.
 
     """
     prepare_work_dir_contents(infile, work_dir)
 
-    DICOMS_PRESENT, DICOMS_SENT = tag_and_transmit.run(
-        work_dir, destination, called_ae, port, calling_ae, group, identifier, tag_value
+    dcms_present, dcms_sent = tag_and_transmit.run(
+        work_dir, destination, called_ae, tls, port, calling_ae, group, identifier, tag_value
     )
-
-    report_generator.generate_report(api_key, parent_acq, infile.name, DICOMS_PRESENT, DICOMS_SENT)
-    
-    
-    
-    return DICOMS_PRESENT, DICOMS_SENT
+    report_generator.generate_report(api_key, parent_acq, infile.name, dcms_present, dcms_sent)
+    return dcms_present, dcms_sent
